+++ packages/components/src/composite/index.ts
@@ -1,4 +1,7 @@
-// Until we migrate away from Reakit, the 'unstable'
-// implementation remains the default.
+// Originally this pointed at a Reakit implementation of
+// `Composite`, but we are removing Reakit entirely from the
+// codebase. We will continue to support the Reakit API
+// through the 'legacy' version, which uses Ariakit under
+// the hood.
 
-export * from './unstable';
+export * from './legacy';